Rough-draft of builders.
I've started working on some "builders", term used loosely.

The goal is to make a sort of service-interface for constructing these
Objects, however I'm not sure what the best way to do this is.

The ultimate goal is to have a VirtualMachineConfigBuilder, which will
either have methods to construct new VMs, modify an existing VM, or
both.

Knowns:

* Closures are good.
* Named-parameters for some of the nullable fields are OK, but I'd like
  to come up with a cleaner way of doing this if possible.
* Want builders to be intuitive and convenient, but still give a lot of
  power and flexibility.

Unknowns:

* Which nullable fields are actually used when creating / modifying the
  components.
* If there are any "mnemonic" ways to do certain builds (i.e. generating
  the ```fileName``` field for the user by having them pass the VM,
  Datastore, and disk label).